move get_all_level_fields to utils, change for usage in both evaluate and annotate_dataset, move tests accordingly as well
